requirements.txt
pymupdf>=1.23.0
python-docx>=1.0.0
odfpy>=1.4.1
sentence-transformers>=2.2.2
qdrant-client>=1.7.0
llama-cpp-python>=0.2.0
torch>=2.0.0
tqdm>=4.65.0
flask>=3.0.0
